\name{homopolymer-package}
\alias{homopolymer-package}
\alias{homopolymer}
\docType{package}
\title{Mathematical Modelling of Homo Polymerization Process}
\description{The package implements the theoretical model described in the references. It deals with the polymer production in a homogeneous phase (bulk, solution)
with batch, semi batch and continuous process. 
}
\details{
\tabular{ll}{
Package: \tab homopolymer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-23\cr
License: \tab GPL-2\cr
}
The package tends to be friendly and  flexible, fully driven by graphical menus. It also has a comprehensive database of physical/chemical characteristics for
the different polymerization ingredients.
Please note : the R-package is made using published information released in the references. It does not use any code of the original authors. So, there is no guarantee the code runs as described in the papers.
 However the package implements the material, energy and micro structural balances exactly as described and uses the optimal datasets highlighted.
}
\author{Gianmarco Polotti}
\references{
\itemize{
	\item{Gao J., Penlidis A. (1996) \emph{A Comprehensive Simulator/Database Package for Reviewing Free-Radical Homopolymerization}, J.M.S.-Rev.Macromol.Chem.Phys, C36(2),199-404}
	\item{Kuindersma M., (1992) \emph{On the Modelling of Free-Radical Polymerization Reactions: Homopolymerization}, M.A.Sc.Thesis, Chemical Engineering, University of Waterloo}
	\item{Gao J., (1992) \emph{Mathematical Modelling of Homopolymerizations: Simulation Package/Database Extensions}, M.A.Sc.Thesis, Chemical Engineering, University of Waterloo}
	\item{Lawrence M.F.,Verzani J., (2012) \emph{Programming Graphical User Interfaces in R},CRC Press Taylor & Francis Group}
}
}
\keyword{ package }
