\name{integral}
\alias{integral}
\title{Homopolymer Integration}
\description{The function integrate the material and energy balance over the time}
\usage{integral(out)}
\arguments{
  \item{out}{a list as given and described by the \code{prepare} function}
}
\details{The function integrate the material and energy balances over the time considering all the physical constrains}
\value{A table of several variable integrate in the required periods, ready for customer visualization and plotting}
\author{Gianmarco Polotti}
\keyword{ homopolymer }
