\name{PlotBasic}
\alias{PlotBasic}
\title{Basic Plots}
\description{Show major trends of polymerization: Conversion, Residual Monomer, Reactor Temperature, Molecular Weight vs. time}
\usage{PlotBasic(out, pars)}
\arguments{
  \item{out}{table of numeric results as given by the function \code{integral}}
  \item{pars}{vector of chemical-physical parameters, as supplied by the function \code{prepare}}
}
\details{Some basic graphics to follow the polymerization}
\value{Just a graphical window with four plots}
\author{Gianmarco Polotti}
\keyword{ homopolymer }
