\name{PlotChoice}
\alias{PlotChoice}
\title{Choose Plot}
\description{User may choose the three quantities to plot in two different scale}
\usage{PlotChoice(out, otab)}
\arguments{
  \item{out}{table of numeric results as given by the function \code{integral}}
  \item{otab}{data.frame with quantities and their descriptions}
}
\details{Some basic graphics to follow the polymerization}
\value{Just a graphical window with a plot}
\author{Gianmarco Polotti}
\keyword{ homopolymer }
