\name{polymer_toolbar}
\alias{polymer_toolbar}
\title{Main toolbar}
\description{Main toolbar for HomoPolymer pacakage}
\usage{polymer_toolbar()}
\details{Just a graphical menu to facilitate the sequence of User instructions}
\value{None. After each command a function is activated and something is printed on the console or in a graphical window.}
\author{Gianmarco Polotti}
\keyword{ HomoPolymer }
\keyword{ toolbar }
